requirements.txt
nltk==3.8.1
numpy==1.24.4
python_dateutil==2.8.2
python_rake==1.5.0
